BH: Rename exaples of dropdown Component on Documentation
